+++ packages/core/provider/src/styles/index.ts
@@ -1,14 +1,20 @@
-import type { ComponentsConfigRequired } from '../types/components'
-import type { GlobalConfigRequired } from '../types/global'
+import type { ComponentSettings } from '../types/components'
+import type { GlobalSettings } from '../types/global'
 import { baseAnimations, baseColors } from './base'
+import { pulse } from './effects/pulse'
 import { ripple } from './effects/ripple'
 import { spin } from './effects/spin'
 
-export const styles = (components: ComponentsConfigRequired, global: GlobalConfigRequired) => {
+export const styles = (
+  components: Required<ComponentSettings>,
+  global: Required<GlobalSettings>
+) => {
   baseColors(global)
   baseAnimations(global)
 
-  spin(global, components.spinner)
+  pulse(global, components.skeleton)
 
   ripple(global)
+
+  spin(global, components.spinner)
 }