+++ packages/ui/data-display/code-block/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @negative-space/code-block
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [96ab878]
+- Updated dependencies [b0697fa]
+  - @negative-space/system@1.5.0
+
 ## 1.0.3
 
 ### Patch Changes