+++ packages/ui/feedback/notification/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @negative-space/notification
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [96ab878]
+- Updated dependencies [b0697fa]
+  - @negative-space/system@1.5.0
+  - @negative-space/alert@1.0.1
+
 ## 1.0.0
 
 ### Major Changes