+++ packages/ui/feedback/progress-bar/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @negative-space/progress-bar
 
+## 1.0.1
+
+### Patch Changes
+
+- 74f6f05: Add aria and role props
+- Updated dependencies [96ab878]
+- Updated dependencies [b0697fa]
+  - @negative-space/system@1.5.0
+
 ## 1.0.0
 
 ### Major Changes